move things around a little to keep the code clean
