+++ src/client/settings/store-tab.ts
@@ -1,285 +1,28 @@
-import { escapeHtml } from "../utils/dom";
 import { jsonHeaders, authHeaders } from "../utils/request";
-import { confirmModal } from "../modules/modals/confirm-modal/confirm";
-import { initLightbox, screenshotUrl } from "./store-lightbox";
-
-const OFFICIAL_REPO_URL =
-  "https://github.com/fccview/fccview-degoog-extensions.git";
-
-interface RepoInfo {
-  url: string;
-  localPath: string;
-  lastFetched: string;
-  name: string;
-  error?: string;
-  repoImage?: string | null;
-}
-
-interface StoreItem {
-  path: string;
-  repoSlug: string;
-  repoUrl: string;
-  repoName: string;
-  name: string;
-  description?: string;
-  version: string;
-  type: "plugin" | "theme" | "engine" | "transport";
-  installed: boolean;
-  installedVersion?: string;
-  updateAvailable?: boolean;
-  screenshots: string[];
-  author?: { name: string; url?: string };
-  pluginType?: string;
-  engineType?: string;
-}
-
-const _normalizeRepoUrl = (url: string): string => {
-  const t = (url || "").trim();
-  return t.endsWith(".git")
-    ? t
-    : t + (t.includes("?") || t.includes("#") ? "" : ".git");
-};
-
-const _formatRelativeTime = (iso: string): string => {
-  try {
-    const d = new Date(iso);
-    const s = Math.round((Date.now() - d.getTime()) / 1000);
-    if (s < 60) return "just now";
-    if (s < 3600) return `${Math.floor(s / 60)} min ago`;
-    if (s < 86400) return `${Math.floor(s / 3600)} hours ago`;
-    return `${Math.floor(s / 86400)} days ago`;
-  } catch {
-    return "";
-  }
-};
-
-function repoImageSrc(repo: RepoInfo, getToken: () => string | null): string {
-  const img = repo.repoImage;
-  if (!img) return "";
-  if (/^https?:\/\//i.test(img)) return img;
-  const token = getToken();
-  const q = token ? `&token=${encodeURIComponent(token)}` : "";
-  return `/api/store/repos/${encodeURIComponent(repo.localPath)}/asset?path=${encodeURIComponent(img)}${q}`;
-}
-
-function pluginTypeLabel(t: string): string {
-  if (t === "command") return "Bang";
-  if (t === "slot") return "Slot";
-  if (t === "search-result-tab") return "Search tab";
-  if (t === "searchBarAction") return "Search bar";
-  return t.charAt(0).toUpperCase() + t.slice(1).replace(/-/g, " ");
-}
-
-function engineTypeLabel(t: string): string {
-  if (t === "web") return "Web";
-  if (t === "images") return "Images";
-  if (t === "videos") return "Videos";
-  if (t === "news") return "News";
-
-  return t.charAt(0).toUpperCase() + t.slice(1);
-}
-
-const _renderRepoDetail = (
-  repo: RepoInfo,
-  getToken: () => string | null,
-  statusByUrl: Record<string, number>,
-): string => {
-  const err = repo.error
-    ? `<span class="store-repo-error">${escapeHtml(repo.error)}</span>`
-    : "";
-  const isOfficial =
-    _normalizeRepoUrl(repo.url) === _normalizeRepoUrl(OFFICIAL_REPO_URL);
-  const removeBtn = isOfficial
-    ? ""
-    : `<button class="btn btn--danger store-btn-remove" type="button" data-url="${escapeHtml(repo.url)}">Remove</button>`;
-  const normUrl = _normalizeRepoUrl(repo.url);
-  const behind = statusByUrl[normUrl] ?? statusByUrl[repo.url] ?? 0;
-  const updatesNote =
-    behind > 0
-      ? `<span class="store-repo-updates-note" title="Refresh to get latest">${escapeHtml(String(behind))} update${behind !== 1 ? "s" : ""} available</span>`
-      : "";
-  const imgSrc = repoImageSrc(repo, getToken);
-  const imgHtml = imgSrc
-    ? `<img src="${escapeHtml(imgSrc)}" alt="" class="store-repo-img" loading="lazy">`
-    : '<div class="store-repo-img store-repo-img-placeholder"></div>';
-  return `
-    <div class="store-repo-detail" data-url="${escapeHtml(repo.url)}">
-      <div class="store-repo-detail-media">${imgHtml}</div>
-      <div class="store-repo-detail-body">
-        <div class="store-repo-name">${escapeHtml(repo.name || repo.url)}</div>
-        <a href="${escapeHtml(repo.url.replace(/\.git$/, ""))}" target="_blank" rel="noopener" class="store-repo-url">${escapeHtml(repo.url)}</a>
-        <div class="store-repo-meta">
-          ${escapeHtml(_formatRelativeTime(repo.lastFetched))}
-          ${err}
-          ${updatesNote}
-        </div>
-        <div class="store-repo-actions">
-          <button class="btn store-btn-refresh" type="button" data-url="${escapeHtml(repo.url)}">Refresh</button>
-          ${removeBtn}
-        </div>
-      </div>
-    </div>`;
-};
-
-const _renderRepoList = (
-  repos: RepoInfo[],
-  getToken: () => string | null,
-  statusByUrl: Record<string, number>,
-  selectedUrl: string | null,
-): string => {
-  if (!repos.length) {
-    return '<p class="store-empty">No repositories added. Add a git repository URL to browse its plugins, themes, engines, and transports.</p>';
-  }
-  const selected = selectedUrl
-    ? repos.find((r) => r.url === selectedUrl)
-    : null;
-  let html = "";
-  html += '<div class="store-repo-list">';
-  for (const repo of repos) {
-    const imgSrc = repoImageSrc(repo, getToken);
-    const active = repo.url === selectedUrl ? " store-repo-item--active" : "";
-    const normUrl = _normalizeRepoUrl(repo.url);
-    const behind = statusByUrl[normUrl] ?? statusByUrl[repo.url] ?? 0;
-    const dot = behind > 0 ? '<span class="store-repo-update-dot"></span>' : "";
-    const imgHtml = imgSrc
-      ? `<img src="${escapeHtml(imgSrc)}" alt="${escapeHtml(repo.name || "")}" class="store-repo-img" loading="lazy">`
-      : '<div class="store-repo-img store-repo-img-placeholder"></div>';
-    html += `
-      <div class="store-repo-item${active}" data-url="${escapeHtml(repo.url)}" role="button" tabindex="0" title="${escapeHtml(repo.name || repo.url)}">
-        <div class="store-repo-item-media">${imgHtml}${dot}</div>
-      </div>`;
-  }
-  html += "</div>";
-  if (selected) {
-    html += _renderRepoDetail(selected, getToken, statusByUrl);
-  }
-  return html;
-};
-
-const _renderItemCard = (
-  item: StoreItem,
-  getToken: () => string | null,
-): string => {
-  const itemSlug = item.path.split("/").pop() ?? "";
-  const token = getToken();
-  const firstUrl = item.screenshots.length
-    ? screenshotUrl(
-        item.repoSlug,
-        item.type,
-        itemSlug,
-        item.screenshots[0],
-        token,
-      )
-    : "";
-  const thumb = item.screenshots.length
-    ? `<img src="${firstUrl}" alt="" class="store-card-thumb" loading="lazy">`
-    : `<div class="store-card-thumb store-card-thumb-placeholder"></div>`;
-  const hasScreenshots = item.screenshots.length > 0;
-  const clickableClass = hasScreenshots
-    ? " store-card-thumb-wrap--clickable"
-    : "";
-  const screenshotsData = hasScreenshots
-    ? ` data-screenshot-files="${escapeHtml(item.screenshots.join(","))}" data-repo-slug="${escapeHtml(item.repoSlug)}" data-item-type="${escapeHtml(item.type)}" data-item-slug="${escapeHtml(itemSlug)}" data-first-screenshot-url="${escapeHtml(firstUrl)}"`
-    : "";
-  const thumbA11y = hasScreenshots
-    ? ' role="button" tabindex="0" aria-label="View screenshots"'
-    : "";
-  const author = item.author
-    ? item.author.url
-      ? `<a href="${escapeHtml(item.author.url)}" target="_blank" rel="noopener">${escapeHtml(item.author.name)}</a>`
-      : escapeHtml(item.author.name)
-    : "";
-
-  let typeLabel = "";
-  let subLabel = "";
-  if (item.type === "plugin") {
-    typeLabel = "Plugin";
-    subLabel = item.pluginType ? pluginTypeLabel(item.pluginType) : "";
-  } else if (item.type === "engine") {
-    typeLabel = "Engine";
-    subLabel = item.engineType ? engineTypeLabel(item.engineType) : "";
-  } else if (item.type === "transport") {
-    typeLabel = "Transport";
-  } else {
-    typeLabel = "Theme";
-  }
-
-  const btn = item.installed
-    ? item.updateAvailable
-      ? `<span class="ext-configured-badge"></span><button class="btn btn--primary store-btn-update" type="button" data-repo-url="${escapeHtml(item.repoUrl)}" data-item-path="${escapeHtml(item.path)}" data-type="${escapeHtml(item.type)}">Update</button><button class="btn btn--secondary store-btn-uninstall" type="button" data-repo-url="${escapeHtml(item.repoUrl)}" data-item-path="${escapeHtml(item.path)}" data-type="${escapeHtml(item.type)}">Uninstall</button>`
-      : `<span class="ext-configured-badge"></span><button class="btn btn--secondary store-btn-uninstall" type="button" data-repo-url="${escapeHtml(item.repoUrl)}" data-item-path="${escapeHtml(item.path)}" data-type="${escapeHtml(item.type)}">Uninstall</button>`
-    : `<button class="btn btn--primary store-btn-install" type="button" data-repo-url="${escapeHtml(item.repoUrl)}" data-item-path="${escapeHtml(item.path)}" data-type="${escapeHtml(item.type)}">Install</button>`;
-  return `
-    <div class="store-card" data-repo-url="${escapeHtml(item.repoUrl)}" data-item-path="${escapeHtml(item.path)}" data-type="${escapeHtml(item.type)}" data-plugin-type="${escapeHtml(item.pluginType || "")}" data-engine-type="${escapeHtml(item.engineType || "")}">
-      <div class="store-card-thumb-wrap${clickableClass}"${screenshotsData}${thumbA11y}>${thumb}</div>
-      <div class="store-card-body">
-        <div class="store-card-main">
-          <div class="store-card-name">${escapeHtml(item.name)}</div>
-          <div class="store-card-meta">by ${author || "—"} · ${escapeHtml(item.repoName)}</div>
-          <div class="store-card-desc">${escapeHtml(item.description || "")}</div>
-          <div class="store-card-version">${item.updateAvailable ? `<span class="store-card-version-old">v${escapeHtml(item.installedVersion || "?")}</span> → ` : ""}v${escapeHtml(item.version)}</div>
-        </div>
-        <div class="store-card-footer">
-          <span class="store-type-badge store-type-${item.type}">${typeLabel}</span>
-          ${subLabel ? `<span class="store-subtype-badge">${escapeHtml(subLabel)}</span>` : ""}
-          <div class="store-card-actions">${btn}</div>
-        </div>
-      </div>
-    </div>`;
-};
-
-const _filterItems = (
-  items: StoreItem[],
-  typeFilter: string,
-  subtypeFilter: string,
-  searchQuery: string,
-  repoFilter: string | null,
-): StoreItem[] => {
-  let out = items;
-  if (repoFilter) {
-    const norm = _normalizeRepoUrl(repoFilter);
-    out = out.filter((i) => _normalizeRepoUrl(i.repoUrl) === norm);
-  }
-  if (typeFilter && typeFilter !== "all") {
-    out = out.filter((i) => i.type === typeFilter);
-  }
-  if (subtypeFilter && subtypeFilter !== "all") {
-    out = out.filter((i) => {
-      if (i.type === "plugin") return i.pluginType === subtypeFilter;
-      if (i.type === "engine") return i.engineType === subtypeFilter;
-      return true;
-    });
-  }
-  if (searchQuery && searchQuery.trim()) {
-    const q = searchQuery.trim().toLowerCase();
-    out = out.filter(
-      (i) =>
-        (i.name && i.name.toLowerCase().includes(q)) ||
-        (i.description && i.description.toLowerCase().includes(q)) ||
-        (i.repoName && i.repoName.toLowerCase().includes(q)) ||
-        (i.author?.name && i.author.name.toLowerCase().includes(q)),
-    );
-  }
-  return out;
-};
-
-function collectSubtypes(items: StoreItem[], typeFilter: string): string[] {
-  if (typeFilter === "plugin") {
-    const set = new Set<string>();
-    items.forEach((i) => {
-      if (i.type === "plugin" && i.pluginType) set.add(i.pluginType);
-    });
-    return Array.from(set).sort();
-  }
-  if (typeFilter === "engine") {
-    const set = new Set<string>();
-    items.forEach((i) => {
-      if (i.type === "engine" && i.engineType) set.add(i.engineType);
-    });
-    return Array.from(set).sort();
-  }
-  return [];
-}
+import { initLightbox } from "./store-lightbox";
+import type { RepoInfo, StoreItem } from "../types/store-tab";
+import { escapeHtml } from "../utils/dom";
+import { getStoreTabHtml } from "./store/store-tab-template";
+import {
+  confirmRemoveRepo,
+  handleAddRepo,
+  handleInstall,
+  handleRefresh,
+  handleRefreshAll,
+  handleRemove,
+  handleUninstall,
+  handleUpdate,
+  handleUpdateAll,
+} from "./store/store-tab-handlers";
+import {
+  collectSubtypes,
+  engineTypeLabel,
+  filterItems,
+  normalizeRepoUrl,
+  pluginTypeLabel,
+  renderItemCard,
+  renderRepoList,
+} from "./store/store-tab-render";
 
 export async function initStoreTab(
   container: HTMLElement,
@@ -316,7 +59,7 @@ export async function initStoreTab(
     const statuses = data.statuses || [];
     const map: Record<string, number> = {};
     for (const s of statuses) {
-      map[_normalizeRepoUrl(s.url)] = s.behind;
+      map[normalizeRepoUrl(s.url)] = s.behind;
       map[s.url] = s.behind;
     }
     repoStatusByUrl = map;
@@ -345,7 +88,7 @@ export async function initStoreTab(
       ".store-repo-list-wrap",
     );
     if (listEl) {
-      listEl.innerHTML = _renderRepoList(
+      listEl.innerHTML = renderRepoList(
         repos,
         getToken,
         repoStatusByUrl,
@@ -436,7 +179,7 @@ export async function initStoreTab(
     }
 
     if (grid) {
-      const filtered = _filterItems(
+      const filtered = filterItems(
         items,
         typeFilter,
         subtypeFilter,
@@ -444,22 +187,31 @@ export async function initStoreTab(
         selectedRepoUrl,
       );
       grid.innerHTML = filtered
-        .map((item) => _renderItemCard(item, getToken))
+        .map((item) => renderItemCard(item, getToken))
         .join("");
       grid
         .querySelectorAll<HTMLButtonElement>(".store-btn-install")
         .forEach((btn) => {
-          btn.addEventListener("click", () => void handleInstall(btn));
+          btn.addEventListener(
+            "click",
+            () => void handleInstall(btn, getToken, loadItems, render),
+          );
         });
       grid
         .querySelectorAll<HTMLButtonElement>(".store-btn-uninstall")
         .forEach((btn) => {
-          btn.addEventListener("click", () => void handleUninstall(btn));
+          btn.addEventListener(
+            "click",
+            () => void handleUninstall(btn, getToken, loadItems, render),
+          );
         });
       grid
         .querySelectorAll<HTMLButtonElement>(".store-btn-update")
         .forEach((btn) => {
-          btn.addEventListener("click", () => void handleUpdate(btn));
+          btn.addEventListener(
+            "click",
+            () => void handleUpdate(btn, getToken, loadItems, render),
+          );
         });
     }
 
@@ -503,229 +255,23 @@ export async function initStoreTab(
           });
         updatesPanel
           .querySelector<HTMLButtonElement>(".store-btn-update-all")
-          ?.addEventListener("click", () => void handleUpdateAll());
+          ?.addEventListener(
+            "click",
+            () => void handleUpdateAll(container, getToken, loadItems, render),
+          );
         updatesPanel
           .querySelectorAll<HTMLButtonElement>(".store-btn-update")
           .forEach((btn) => {
-            btn.addEventListener("click", () => void handleUpdate(btn));
+            btn.addEventListener(
+              "click",
+              () => void handleUpdate(btn, getToken, loadItems, render),
+            );
           });
       }
     }
   }
 
-  function showError(el: HTMLElement | null, msg: string): void {
-    if (!el) return;
-    el.textContent = msg;
-    el.classList.add("store-error-visible");
-    setTimeout(() => el.classList.remove("store-error-visible"), 4000);
-  }
-
-  async function handleAddRepo(
-    inputEl: HTMLInputElement | null,
-    addBtn: HTMLButtonElement,
-    errorEl: HTMLElement | null,
-  ): Promise<void> {
-    const url = inputEl?.value?.trim();
-    if (!url) return;
-    addBtn.disabled = true;
-    if (errorEl) errorEl.textContent = "";
-    try {
-      const res = await fetch("/api/store/repos", {
-        method: "POST",
-        headers: jsonHeaders(getToken),
-        body: JSON.stringify({ url }),
-      });
-      const data = (await res.json()) as { error?: string };
-      if (!res.ok) {
-        showError(errorEl, data.error || "Failed to add repository");
-        return;
-      }
-      if (inputEl) inputEl.value = "";
-      await refreshAndRender();
-    } catch {
-      showError(errorEl, "Network error");
-    } finally {
-      addBtn.disabled = false;
-    }
-  }
-
-  async function handleRefresh(url: string): Promise<void> {
-    const res = await fetch("/api/store/repos/refresh", {
-      method: "POST",
-      headers: jsonHeaders(getToken),
-      body: JSON.stringify({ url }),
-    });
-    if (!res.ok) return;
-    await refreshAndRender();
-    void loadReposStatus().then(() => render());
-  }
-
-  async function handleRemove(url: string): Promise<void> {
-    const fromRepo = repos.find((r) => r.url === url);
-    if (!fromRepo) return;
-    const res = await fetch("/api/store/repos", {
-      method: "DELETE",
-      headers: jsonHeaders(getToken),
-      body: JSON.stringify({ url }),
-    });
-    const data = (await res.json()) as { error?: string };
-    if (!res.ok) {
-      alert(data.error || "Failed to remove repository");
-      return;
-    }
-    await refreshAndRender();
-  }
-
-  async function handleInstall(btn: HTMLButtonElement): Promise<void> {
-    const { repoUrl, itemPath, type } = btn.dataset;
-    if (
-      type === "plugin" &&
-      !(await confirmModal({
-        title: "Install plugin?",
-        message:
-          "This plugin will run code on your server. Only install from sources you trust. Continue?",
-      }))
-    )
-      return;
-    btn.disabled = true;
-    try {
-      const res = await fetch("/api/store/install", {
-        method: "POST",
-        headers: jsonHeaders(getToken),
-        body: JSON.stringify({ repoUrl, itemPath, type }),
-      });
-      const data = (await res.json()) as { error?: string };
-      if (!res.ok) alert(data.error || "Install failed");
-      else {
-        await loadItems();
-        render();
-        window.dispatchEvent(new CustomEvent("extensions-saved"));
-      }
-    } catch {
-      alert("Network error");
-    } finally {
-      btn.disabled = false;
-    }
-  }
-
-  async function handleUninstall(btn: HTMLButtonElement): Promise<void> {
-    const { repoUrl, itemPath, type } = btn.dataset;
-    if (
-      !(await confirmModal({
-        title: "Uninstall?",
-        message: `Uninstall this ${type ?? "item"}?`,
-      }))
-    )
-      return;
-    btn.disabled = true;
-    try {
-      const res = await fetch("/api/store/uninstall", {
-        method: "POST",
-        headers: jsonHeaders(getToken),
-        body: JSON.stringify({ repoUrl, itemPath, type }),
-      });
-      const data = (await res.json()) as { error?: string };
-      if (!res.ok) alert(data.error || "Uninstall failed");
-      else {
-        await loadItems();
-        render();
-        window.dispatchEvent(new CustomEvent("extensions-saved"));
-      }
-    } catch {
-      alert("Network error");
-    } finally {
-      btn.disabled = false;
-    }
-  }
-
-  async function handleUpdate(btn: HTMLButtonElement): Promise<void> {
-    const { repoUrl, itemPath, type } = btn.dataset;
-    btn.disabled = true;
-    try {
-      const res = await fetch("/api/store/update", {
-        method: "POST",
-        headers: jsonHeaders(getToken),
-        body: JSON.stringify({ repoUrl, itemPath, type }),
-      });
-      const data = (await res.json()) as { error?: string };
-      if (!res.ok) alert(data.error || "Update failed");
-      else {
-        await loadItems();
-        render();
-        window.dispatchEvent(new CustomEvent("extensions-saved"));
-      }
-    } catch {
-      alert("Network error");
-    } finally {
-      btn.disabled = false;
-    }
-  }
-
-  async function handleUpdateAll(): Promise<void> {
-    const btn = container.querySelector<HTMLButtonElement>(
-      ".store-btn-update-all",
-    );
-    if (btn) btn.disabled = true;
-    try {
-      const res = await fetch("/api/store/update-all", {
-        method: "POST",
-        headers: jsonHeaders(getToken),
-      });
-      const data = (await res.json()) as { error?: string; updated?: number };
-      if (!res.ok) alert(data.error || "Update failed");
-      else {
-        await loadItems();
-        render();
-        window.dispatchEvent(new CustomEvent("extensions-saved"));
-      }
-    } catch {
-      alert("Network error");
-    } finally {
-      if (btn) btn.disabled = false;
-    }
-  }
-
-  container.innerHTML = `
-    <section class="store-repos-section settings-section">
-      <div class="store-repos-header">
-        <h2 class="settings-section-heading">Repositories</h2>
-        <div class="header-actions">
-          <div class="store-repos-actions">
-            <button class="btn store-btn-refresh-all" type="button">Refresh all</button>
-          </div>
-          <button class="btn btn--primary store-btn-add" type="button">Add repository</button>
-        </div>
-      </div>
-      <div class="store-add-repo-wrap" style="display:none">
-        <input type="text" class="store-input-url" placeholder="https://github.com/user/repo.git">
-        <button class="btn btn--primary store-btn-add-confirm" type="button">Add</button>
-        <span class="store-inline-error"></span>
-      </div>
-      <p class="settings-desc">Add a git repository URL to browse and install plugins, themes, engines, and transports. Set <code>repo-image</code> in the repo’s package.json to show an image next to the URL.</p>
-      <div class="store-repo-list-wrap"></div>
-    </section>
-    <section class="store-catalog-section settings-section">
-      <div class="store-catalog-header">
-        <h2 class="settings-section-heading">Catalog</h2>
-      </div>
-      <div class="store-updates-panel" style="display:none"></div>
-      <div class="store-filter-bar">
-        <input type="text" class="store-search-input" placeholder="Search…" id="store-search-input">
-        <select class="store-filter-select store-filter-type" aria-label="Filter by type"></select>
-        <select class="store-filter-select store-filter-subtype" aria-label="Filter by sub-type" style="display:none"></select>
-      </div>
-      <div class="store-catalog-grid"></div>
-    </section>
-    <div class="store-lightbox" id="store-lightbox" aria-hidden="true" role="dialog" aria-modal="true" aria-label="Screenshot gallery">
-      <div class="store-lightbox-backdrop"></div>
-      <button class="store-lightbox-close" type="button" aria-label="Close">&times;</button>
-      <button class="store-lightbox-prev" type="button" aria-label="Previous">&larr;</button>
-      <div class="store-lightbox-img-wrap">
-        <img class="store-lightbox-img" src="" alt="">
-      </div>
-      <button class="store-lightbox-next" type="button" aria-label="Next">&rarr;</button>
-      <div class="store-lightbox-counter"></div>
-    </div>`;
+  container.innerHTML = getStoreTabHtml();
 
   initLightbox(container, getToken);
 
@@ -746,27 +292,26 @@ export async function initStoreTab(
         addWrap.style.display === "none" ? "block" : "none";
   });
   addConfirmBtn?.addEventListener("click", () => {
-    if (addConfirmBtn) void handleAddRepo(urlInput, addConfirmBtn, addErrorEl);
+    if (addConfirmBtn)
+      void handleAddRepo(
+        urlInput,
+        addConfirmBtn,
+        addErrorEl,
+        getToken,
+        refreshAndRender,
+      );
   });
 
   container
     .querySelector<HTMLButtonElement>(".store-btn-refresh-all")
     ?.addEventListener("click", async () => {
-      const btn = container.querySelector<HTMLButtonElement>(
-        ".store-btn-refresh-all",
+      await handleRefreshAll(
+        container,
+        getToken,
+        refreshAndRender,
+        loadReposStatus,
+        render,
       );
-      if (btn) btn.disabled = true;
-      try {
-        await fetch("/api/store/repos/refresh", {
-          method: "POST",
-          headers: jsonHeaders(getToken),
-          body: JSON.stringify({}),
-        });
-        await refreshAndRender();
-        void loadReposStatus().then(() => render());
-      } finally {
-        if (btn) btn.disabled = false;
-      }
     });
 
   container.addEventListener("click", async (e) => {
@@ -776,14 +321,23 @@ export async function initStoreTab(
     const removeBtn = (e.target as HTMLElement).closest<HTMLElement>(
       ".store-btn-remove",
     );
-    if (refreshBtn?.dataset.url) void handleRefresh(refreshBtn.dataset.url);
+    if (refreshBtn?.dataset.url)
+      void handleRefresh(
+        refreshBtn.dataset.url,
+        getToken,
+        refreshAndRender,
+        loadReposStatus,
+        render,
+      );
     if (removeBtn?.dataset.url) {
-      const ok = await confirmModal({
-        title: "Remove repository?",
-        message:
-          "Remove this repository? You must uninstall any installed items first.",
-      });
-      if (ok) void handleRemove(removeBtn.dataset.url);
+      const ok = await confirmRemoveRepo(removeBtn.dataset.url);
+      if (ok)
+        void handleRemove(
+          removeBtn.dataset.url,
+          repos,
+          getToken,
+          refreshAndRender,
+        );
     }
   });
